Struct cargo_manifest::Package

pub struct Package<Metadata = Value> {

    pub name: String,
    pub edition: Edition,
    pub version: String,
    pub build: Option<Value>,
    pub workspace: Option<String>,
    pub authors: Vec<String>,
    pub links: Option<String>,
    pub description: Option<String>,
    pub homepage: Option<String>,
    pub documentation: Option<String>,
    pub readme: Option<StringOrBool>,
    pub keywords: Vec<String>,
    pub categories: Vec<String>,
    pub license: Option<String>,
    pub license_file: Option<String>,
    pub repository: Option<String>,
    pub metadata: Option<Metadata>,
    pub default_run: Option<String>,
    pub autobins: bool,
    pub autoexamples: bool,
    pub autotests: bool,
    pub autobenches: bool,
    pub publish: Publish,
    pub resolver: Option<Resolver>,

}

You can replace Metadata type with your own to parse into something more useful than a generic toml Value

Fields
name: String

Careful: some names are uppercase

edition: Editionversion: String

e.g. “1.9.0”

build: Option<Value>workspace: Option<String>authors: Vec<String>

e.g. [“Author e@mail”, “etc”]

links: Option<String>description: Option<String>

A short blurb about the package. This is not rendered in any format when uploaded to crates.io (aka this is not markdown).

homepage: Option<String>documentation: Option<String>readme: Option<StringOrBool>

This points to a file under the package root (relative to this Cargo.toml).

keywords: Vec<String>categories: Vec<String>

This is a list of up to five categories where this crate would fit. e.g. [“command-line-utilities”, “development-tools::cargo-plugins”]

license: Option<String>

e.g. “MIT”

license_file: Option<String>repository: Option<String>metadata: Option<Metadata>default_run: Option<String>

The default binary to run by cargo run.

autobins: boolautoexamples: boolautotests: boolautobenches: boolpublish: Publishresolver: Option<Resolver>